Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Just to complement the comment:

In Spare Parts the ON and OFF summary where equal for this preference, which made no sense and when user tapped the checkbox the text would not change accordingly.

The preference name is "Haptic feedback".

Patch-set: 1