Update patch set 1
Patch Set 1:

I have a xhdpi icon done too :) Only problem is the way it's setup, looks like a Honeycomb thing? With "mipmap-hdpi", is it possible to add a mipmap-xhdpi? I was honestly thinking of overhauling the whole app and merging api tools into dev tools. For now this works great and looks good on xhdpi devices. This is my first commit so I didn't want to over do it ;)

Patch-set: 1