Update patch set 1
Patch Set 1:

Yes, creating a folder namned mipmap-xhdpi in this case and putting the correctly sized and properly named icon inside it will make sure the correct icon gets loaded by Android on xhdpi-devices.

Perhaps the icon should be more similar to the original, ie. include one or more gears but with the proper perspective and following the ICS style guidelines.

Patch-set: 1