Update patch set 1
Patch Set 1:

This new icon follows the ICS UI guidelines :) I can make a new one with gears thats not a problem at all. I'm personally sick of gears though. Gears in Rom Manager, Spare Parts, Dev Tools, API Tools, and a few more. Regardless if this one is used or not, the old need to be replaced to match ICS. Is there something I can change on the new one so everyone is happy?

Patch-set: 1