Update patch set 1
Patch Set 1:

> Upload to AOSP review as well?

I didn't intend to, but I will if it is encouraged/suggested. 
I saw on wiki that build instructions are macOS friendly,
and indeed build env works relatively without a flaw on 
Sierra, so a couple of issues I tackled along the way, 
such as the one regarding generating own key-sets 
for signing personal builds, I'd be glad to share back.

Patch-set: 1