cargo_embargo: Allow no_std variant to suppress props
Empty lists previously would show up as "no override" when merging
modules. If the std module has one of these lists populated, and the
no_std module has it empty, maintaining the behavior of the no_std
module requires it to override it to be empty for the no_std variant
explicitly.

Bug: 446969268
Test: Regenerate no_std split of thiserror
Change-Id: I9a7bc2cd4cdffabcea5bc74c9c545147bae2bdc2