Compatibilize code with google3
```
no-inferable-primitives: Type boolean is trivially inferred from a boolean literal, remove type annotation.
```

```
Exported symbols must have JSDoc comments (see go/tsstyle#comments-documentation). Explain why this code exists, and what it's useful for. If this was triggered on a pattern that doesn't need a comment, an exception can be added in the enforce-comments-on-exported-symbols tslint rule.
```

Bug: 311642700
Test: npm run test:unit:ci
Flag: EXEMPT EXEMPT
Change-Id: I0b181f161d7511152f3d5e200881528ca233a458