Add default log_from to group_overrides
When using DEFAULT tracing_mode in the ProtoLog config, each group_override must specify a `log_from` level to enable tracing for that group. Without it, no log level is enabled, and no data is collected.

This change adds a default `log_from` value to ensure that when a user selects a specific ProtoLog group for tracing, the traces are collected as expected.

Fixes: b/450230245
Test: manually collect a trace on winscope with a protolog group selected
Change-Id: Ib884cea1d5f739c77d5cb840db52eaacc148f986