STL: Default GestureContext from intrinsicDirection 2/2
## Add intrinsicDirection hint to transition DSL

This property acts as a high-level hint for the overall direction of a
transition. It is intended to be used by triggered animations as a
fallback for `MotionValue`s that require a direction but do not have one
explicitly defined.

## Default GestureContext for Transition

This new implementation provides a `GestureContext` if and only if
`transformationSpec.intrinsicDirection` is defined in the transition's
DSL. This allows programmatic animations (like `animateToScene`) to
provide directional context to mechanics, which they previously could
not.

A key consequence is that `gestureContext != null` is no longer a
reliable indicator of a user-driven gesture. Consumers must now use
`isInitiatedByUserInput` for that check (as seen in BouncerContent).

Test: atest VerticalTactileSurfaceRevealModifierTest
Bug: 392535471
Flag: com.android.systemui.qs_tiles_reveal_animation
Change-Id: Iefef1a64cbf0b7e0ab85ef4c28794ade0359ff2a