Provide a list of targeted-crates for user to update in the generated email.
Stalled crate updates are tracked in a buganizer hotlist. This change will Fetch the current list of stalled crates from the hotlist using the bugged tool and filter them out when suggesting crates. The suggest list of crates will be added to an email that is generated when running the automatic updater.
bug: b/442878732
bug: b/442880639
Test: cargo run  -- ~/source/main-without-vendor
Change-Id: Id31d0fa195fc9d68ba1468df0b46c5192a3fffc5