Encode device id as URI component.
Version patch increment for compatibility.

Old version of Winscope app will work with new proxy, as the non-encoded device ids will be unchanged by urllib#parse. New version of Winscope will not work with old proxy, as it sends encoded device ids which are not parsed by the old proxy.

Hence, a patch increment is necessary, so app 6.0.2 will register as incompatible with script 6.0.1 but app 6.0.1 will be compatible with script 6.0.2.

Bug: 455757838
Test: manual
Change-Id: I7c856a2d44734c621f90a5c0fe76279f7322a63d