license_checker: Apply fuzzy checker to remnants of license file
If a crate supplies a single license file which contains multiple
licenses, currently we only succeed at identifying them if all licenses
are character-for-character identical. Little things like pluralization
can throw it off, which is what the fuzzy checking pass was intended
for. However, the fuzzy check only ran if the exact check failed, and
could only produce a single license.

Instead, run the fuzzy check on the remnants after the exact check is
done. This is still somewhat limited in that it will only retrieve a
single fuzzy license, but it will still find that license in a file with
multiple other exact licenses.

Bug: 459897955
Test: Successfully regenerate aws-lc-rs
Change-Id: I5013c038921769895c125241b3f25ae99d60414e