Introduce UiTreeNode and UiTreeNodeRow.
At the presenter level we will flatten a tree in DFS order into a list
of rows. This allows us to render only a subset of rows depending on
what should be visible in the viewport, via a custom alternative
implementation to the Angular virtual scroll viewport.

Since we will be processing a tree into a DFS traversal list of rows we
no longer need to maintain DFS-order references between nodes in a tree.

Bug: 440637831
Test: npm run test:unit:ci
Change-Id: I1a87a2f8fc5a7e689f354f18824457210b80c5b4