Add command to approximate how many update commits were submitted over time.
This command will examine the results of running `git log`. It will include every commit with a message that uses the word "Update" in the first line. It will not include commits that include the word Merge or Revert. It will total the number of commits based on year and quarter. Note the number of commit isn't an exact match to number of updates, since some commits include updates to multiple crates at once. This is just mean to provide an approximation to let us see trends.

Bug: b/442878732
Test: ./crate_tool count-updates
Change-Id: I80352717ddd3ea91d0fdfb63a2a2f9056e9e495b