Disallow updates from regressing crates
Provides a check for the version of the crate provided by the user
and the version of the crate we have in Android. It will panic if the provided version is older than the current crate. This will prevent mistakes.  It will allow the user to regress the crate with use of the flag "--allow_older".

Bug: b/469136918
Bug: b/442878732
Test: ./crate_tool  update dashmap 1.5.3 (blocked because it's younger than dashmap current version v5.5.3)
Test: ./crate_tool  update dashmap 5.15.3 (passes and looks for it on crates.io)
Test: ./crate_tool  update annotate-snippets 0.9.5 (does not pass. Current version 0.11.5)

Change-Id: Ifccae91ccc47f400f0280ecffa5124b77582a244